Finish RC1 for loading existing rooms
